Update to 64bit, not completed

+++ Sources/Elastos/Runtime/Core/reflection/CClsModule.h
@@ -46,7 +46,7 @@ public:
 public:
     CLSModule*      mClsMod;
     Boolean         mAllocedClsMode;
-    Int32           mBase;
+    uintptr_t       mBase;
 
 private:
     TypeAliasDesc*  mTypeAliasList;
