+++ Sources/ToolChains/Carl/ElCar/Lube/lubc/token.cpp
@@ -14,6 +14,7 @@
 // limitations under the License.
 //=========================================================================
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -58,7 +59,7 @@ PVOID SaveTokenContext()
 
 void RestoreTokenContext(PVOID pvCtx)
 {
-    DWORD dwContext = (DWORD)pvCtx;
+    DWORD dwContext = (DWORD)(uintptr_t)pvCtx;
 
     if (dwContext & 0x40000000) s_bBrace = true;
     if (dwContext & 0x80000000) s_bIn = true;
